[BE] add: interface 추가와 그에 따른 클래스 이름 변경
